All tests passing with wire protocol 13
