+++ src/ibase/srp.rs
@@ -63,7 +63,24 @@ use std::{error, fmt, marker::PhantomData};
 
 use digest::Digest;
 use generic_array::GenericArray;
+use lazy_static::lazy_static;
 use num_bigint::BigUint;
+
+lazy_static! {
+    /// Srp Group used by the firebird server
+    pub static ref SRP_GROUP: SrpGroup = SrpGroup {
+        n: BigUint::from_bytes_be(&[
+            230, 125, 46, 153, 75, 47, 144, 12, 63, 65, 240, 143, 91, 178, 98, 126, 208, 212, 158,
+            225, 254, 118, 122, 82, 239, 205, 86, 92, 214, 231, 104, 129, 44, 62, 30, 156, 232,
+            240, 168, 190, 166, 203, 19, 205, 41, 221, 235, 247, 169, 109, 74, 147, 181, 93, 72,
+            141, 240, 153, 161, 92, 137, 220, 176, 100, 7, 56, 235, 44, 189, 217, 168, 247, 186,
+            181, 97, 171, 27, 13, 193, 198, 205, 171, 243, 3, 38, 74, 8, 209, 188, 169, 50, 209,
+            241, 238, 66, 139, 97, 157, 151, 15, 52, 42, 186, 154, 101, 121, 59, 139, 47, 4, 26,
+            229, 54, 67, 80, 193, 111, 115, 95, 86, 236, 188, 168, 123, 213, 123, 41, 231,
+        ]),
+        g: BigUint::from_bytes_be(&[2]),
+    };
+}
 
 /// SRP client state before handshake with the server.
 pub struct SrpClient<'a, D: Digest> {
@@ -272,5 +289,81 @@ impl SrpGroup {
         buf[l..].copy_from_slice(&g_bytes);
 
         BigUint::from_bytes_be(&D::new().chain(&n).chain(&buf).finalize())
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::SRP_GROUP;
+    use crate::ibase::srp::{srp_private_key, SrpClient};
+    use num_bigint::BigUint;
+    use sha1::Sha1;
+
+    #[test]
+    fn srp_group_k_test() {
+        use sha1::Digest;
+
+        let k = {
+            let n = SRP_GROUP.n.to_bytes_be();
+            let g_bytes = SRP_GROUP.g.to_bytes_be();
+            let mut buf = vec![0u8; n.len()];
+            let l = n.len() - g_bytes.len();
+            buf[l..].copy_from_slice(&g_bytes);
+
+            BigUint::from_bytes_be(&sha1::Sha1::new().chain(&n).chain(&buf).finalize())
+        };
+
+        assert_eq!(
+            "1277432915985975349439481660349303019122249719989",
+            &k.to_string()
+        );
+    }
+
+    #[test]
+    fn srp_vals_test() {
+        let user = b"sysdba";
+        let password = b"masterkey";
+
+        // Real one randomly generated
+        let seed = [
+            104, 168, 26, 157, 227, 194, 41, 70, 204, 234, 48, 50, 217, 147, 39, 186, 223, 61, 125,
+            154, 223, 9, 54, 220, 163, 109, 222, 183, 78, 242, 217, 218,
+        ];
+
+        let cli = SrpClient::<Sha1>::new(&seed, &SRP_GROUP);
+
+        assert_eq!(
+            cli.get_a_pub(),
+            BigUint::parse_bytes(
+                b"140881421499567234926370707691929201584335514055692587180102084646282810733160001237892692305806957785292091467614922078328787082920091583399296847456481914076730273969778307678896596634071762017513173403243965936903761580099023780256639030075360658492420403842461445358536578442895018174380364815053686107255"
+                , 10
+            ).unwrap().to_bytes_be()
+        );
+
+        // Real ones are received from server
+        let salt = b"9\xe0\xee\x06\xa9]\xbe\xa7\xe4V\x08\xb1g\xa1\x93\x19\xf6\x11\xcb@\t\xeb\x9c\xf8\xe5K_;\xd1\xeb\x0f\xde";
+        let serv_pub = BigUint::parse_bytes(
+            b"9664511961170061978805668776377548609867359536792555459451373100540811860853826881772164535593386333263225393199902079347793807335504376938377762257920751005873533468177562614066508611409115917792525726727162676806787115902775303095022305576987173568527110065130456437265884455358297687922316181717357090556", 
+            10
+        ).unwrap().to_bytes_be();
+
+        let cli_priv = srp_private_key::<Sha1>(user, password, salt);
+
+        assert_eq!(
+            b"\xe7\xd1>*\xaag\x9a\xa9\"w\x17&>\xca\xff\x86+ '\xdc",
+            &cli_priv[..]
+        );
+
+        let verifier = cli.process_reply(user, salt, &cli_priv, &serv_pub).unwrap();
+
+        assert_eq!(
+            b"C~\xe6\xad\xe1\x97d\xed\xbf\x16D7\xb1C\xbf\xb1\xc9\x92\xc4@",
+            &verifier.get_proof()[..]
+        );
+
+        assert_eq!(
+            b"\xd5,\xe6(\xf6\x04\xec\xdb\xf2\xa2J\xc8zw\xb0\x9a\x87O\xe8\xf7",
+            &verifier.get_key()[..]
+        );
     }
 }
